WIP: Use drm_fbdev_generic_setup to install a framebuffer console
Enable by setting the `modeset=1` and `fbdev=1` parameters.

WARNING: This change is a work in progress and the functionality it
provides is experimental. Use at your own risk.

This framebuffer console support works with drm-kms clients such as
Wayland compositors, but does not interact correctly with native
nvidia-modeset clients such as the NVIDIA X driver or Vulkan
direct-to-display.